chore: uri -> SelectedMedia로 변환하는 함수를 util 폴더에 별도파일로 분리
