+++ android/core/ui/src/main/kotlin/com/andlife/ui/component/invitation/InvitationMediaUpload.kt
@@ -1,8 +1,5 @@
 package com.andlife.ui.component.invitation
 
-import android.content.Context
-import android.media.MediaMetadataRetriever
-import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Arrangement
@@ -36,6 +33,7 @@ import com.andlife.designsystem.theme.InvitationSpacing
 import com.andlife.designsystem.theme.InvitationTheme
 import com.andlife.ui.R
 import com.andlife.ui.model.UiMediaType
+import com.andlife.ui.util.media.uriToSelectedMedia
 
 @Composable
 fun InvitationMediaUpload(
@@ -226,49 +224,4 @@ private fun InvitationMediaUploadPreview() {
             onMediaRemove = {},
         )
     }
-}
-
-private fun uriToSelectedMedia(
-    context: Context,
-    uriString: String,
-): SelectedMedia {
-    val uri = Uri.parse(uriString)
-
-    val mimeType =
-        try {
-            context.contentResolver.getType(uri)
-        } catch (e: Exception) {
-            null
-        }
-
-    val mediaType =
-        when {
-            mimeType?.startsWith("image/") == true -> UiMediaType.IMAGE
-            mimeType?.startsWith("video/") == true -> UiMediaType.VIDEO
-            mimeType?.startsWith("audio/") == true -> UiMediaType.AUDIO
-            else -> UiMediaType.IMAGE
-        }
-
-    val duration =
-        if (mediaType == UiMediaType.VIDEO || mediaType == UiMediaType.AUDIO) {
-            try {
-                val retriever = MediaMetadataRetriever()
-                retriever.setDataSource(context, uri)
-                retriever
-                    .extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION)
-                    ?.toLongOrNull()
-                    ?.div(1000)
-                    ?.toInt()
-            } catch (e: Exception) {
-                null
-            }
-        } else {
-            null
-        }
-
-    return SelectedMedia(
-        uri = uriString,
-        type = mediaType,
-        duration = duration,
-    )
 }
